% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.y}
\alias{data.y}
\title{A random vector generated from normal distribution}
\format{
A numeric vector of length 100
\describe{
\item{V1}{...}
\item{V2}{...}
...
}
}
\usage{
data.y
}
\description{
Some description of the data
}
\keyword{datasets}
